Decommission only registered brokers in CC during downscale
When downscale is initiated for brokers that are not yet part of Kafka cluster
we avoid now decommissioning the broker in CC (i.e. `remove_broker`) to avoid deadlock (CC returns 5XX when unknown brokers are removed)

Also - downscaling brokers not fully registered may miss ConfigMap
so we skip missing ConfigMaps in that case.