Expose http healthcheck endpoint to envoy anycast ingress
When envoy is selected as an ingress for kafka,
a new endpoint :8080/healthcheck is exposed by each envoy
pod and in turn available through the Envoy Service.
This is implemented as an active heathcheck in envoy configured to probe jmx exporter /-/healthy endpoint for each broker.

This is an http port that can be probed to verify
if there is at least one broker up in cluster.
In turn, this can be used in fronting LoadBalancers
that support configuring an http healthcheck port.

Signed-off-by: Adi Muraru <amuraru@adobe.com>