chore(go-mod): removed local replacements
Previously the local sub-Go-modules (api,
properties) were depended upon by the root Go
module using a local replacement.

Removed this because it is not transitive to
downstream usage - where Koperator is also a
dependency - and is anyway only neede for
development purposes which can be kept on local
machine instead of having it committed to the
repo.